[NFC] Moving some code from headers to designated .cpp files. Renaming some files. Extending .gitignore.
