+++ frontend/ast/Type.hpp
@@ -37,55 +37,7 @@ public:
 
   bool IsPointerType() const { return PointerLevel != 0; }
 
-  static std::string ToString(const Type *t) {
-    std::string Result = "";
-    
-    switch (t->GetTypeVariant()) {
-    case Double:
-      Result = "double";
-      break;
-    case Char:
-      Result = "char";
-      break;
-    case UnsignedChar:
-      Result = "unsigned char";
-      break;
-    case Int:
-      Result = "int";
-      break;
-    case UnsignedInt:
-      Result = "unsigned int";
-      break;
-    case Long:
-      Result = "long";
-      break;
-    case UnsignedLong:
-      Result = "unsigned long";
-      break;
-    case LongLong:
-      Result = "long long";
-      break;
-    case UnsignedLongLong:
-      Result = "unsigned long long";
-      break;
-    case Void:
-      Result = "void";
-      break;
-    case Composite:
-      Result = t->GetName();
-      break;
-    case Invalid:
-      return "invalid";
-    default:
-      assert(!"Unknown type.");
-      break;
-    }
-
-    for (size_t i = 0; i < t->GetPointerLevel(); i++)
-      Result.push_back('*');
-
-    return Result;
-  }
+  static std::string ToString(const Type *t);
 
   /// Given two type variants it return the stronger one.
   /// Type variants must be numerical ones.
@@ -299,30 +251,7 @@ public:
       return !(lhs == rhs);
   }
 
-  std::string ToString() const {
-    if (IsFunction()) {
-      auto TyStr = Type::ToString(this);
-      auto ArgSize = ParameterList.size();
-      if (ArgSize > 0)
-        TyStr += " (";
-      for (size_t i = 0; i < ArgSize; i++) {
-        TyStr += Type::ToString(&ParameterList[i]);
-        if (i + 1 < ArgSize)
-          TyStr += ",";
-        else
-          TyStr += ")";
-      }
-      return TyStr;
-    } else if (Kind == Array) {
-      auto TyStr = Type::ToString(this);
-
-      for (size_t i = 0; i < Dimensions.size(); i++)
-        TyStr += "[" + std::to_string(Dimensions[i]) + "]";
-      return TyStr;
-    } else {
-      return Type::ToString(this);
-    }
-  }
+  std::string ToString() const;
 
 private:
   std::string Name; // For structs
