+++ frontend/parser/SymbolTable.hpp
@@ -39,34 +39,11 @@ public:
 
   void InsertGlobalEntry(const Entry &e) { SymTabStack[0].push_back(e); }
 
-  bool Contains(Entry e) {
-    for (int i = Size() - 1; i >= 0; i--) {
-      auto table = SymTabStack[i];
-      for (int j = table.size() - 1; j >= 0; j--)
-        if (e == table[j])
-          return true;
-    }
-    return false;
-  }
+  bool Contains(Entry e);
 
-  std::optional<Entry> Contains(const std::string &sym) {
-    for (int i = Size() - 1; i >= 0; i--) {
-      auto table = SymTabStack[i];
-      for (int j = table.size() - 1; j >= 0; j--)
-        if (sym == std::get<0>(table[j]))
-          return table[j];
-    }
-    return std::nullopt;
-  }
+  std::optional<Entry> Contains(const std::string &sym);
 
-  bool ContainsInCurrentScope(Entry e) {
-    auto idx = Size() > 0 ? Size() - 1 : 0;
-
-    for (int i = SymTabStack[idx].size() - 1; i >= 0; i--)
-      if (e == SymTabStack[idx][i])
-        return true;
-    return false;
-  }
+  bool ContainsInCurrentScope(Entry e);
 
 private:
   std::vector<Table> SymTabStack;
